FEATURE: PLUGINS
This feature for the `IAmACommand` class will allow users of the gem to
create custom plugins to inject new behaviors.  Not only will this give each
user the ability to tailor Commando commands to their exact needs, but
also this will allow the project to remove the dependencies on `Virtus` and
`ActiveModel` and move those to separate plugin gems.